Fix warnings associated with 64-bit/32-bit compiler.

+++ Categories/NSNumber+Buckets.m
@@ -20,16 +20,16 @@
     NSUInteger intVal = [self unsignedIntegerValue];
 
     if( intVal < bucketSize )
-        return [NSString stringWithFormat:@"< %i", bucketSize];
+        return [NSString stringWithFormat:@"< %lu", (unsigned long)bucketSize];
 
     NSUInteger multiple = intVal / bucketSize;
 
     if( multiple >= maxBuckets )
-        return [NSString stringWithFormat:@">= %i", ( bucketSize * maxBuckets )];
+        return [NSString stringWithFormat:@">= %lu", (unsigned long)(bucketSize * maxBuckets )];
     
-    return [NSString stringWithFormat:@"%i - %i",
-		    ( multiple * bucketSize ),
-		    ( ( ( multiple + 1 ) * bucketSize ) - 1)];
+    return [NSString stringWithFormat:@"%lu - %lu",
+		    (unsigned long)(multiple * bucketSize ),
+		    (unsigned long)(((multiple + 1) * bucketSize ) - 1)];
 }
 
 @end
